Fixed problem where sleeping for 0 seconds in the publisher task switches context faster than what allows the NFCReaderDevice to read tags in a separate thread.
Added basic logging in reader.py

Fixed incorrect checking if user is recorded in quest or not (boolean was inverted)

Changed beep method in NFCReaderDevice to a more customizable method where led and buzzer can be controlled  descriptively.

Removed open method in NFCReaderDevice and made it so the device is opened when initialising the object

When initializing the NFCReaderDevice, the nfc reader now buzzes and blinks in a unique pattern to indicated that it is on.